+++ src/cpp/opportunisticsecuresmtpclient.cpp
@@ -1,15 +1,10 @@
 #include "opportunisticsecuresmtpclient.hpp"
 #include <string>
-#include <utility>
 
 using namespace jed_utils::cpp;
 
 OpportunisticSecureSMTPClient::OpportunisticSecureSMTPClient(const std::string &pServerName, unsigned int pPort)
     : jed_utils::OpportunisticSecureSMTPClient(pServerName.c_str(), pPort) {
-}
-
-OpportunisticSecureSMTPClient::~OpportunisticSecureSMTPClient() {
-    delete mCredential;
 }
 
 std::string OpportunisticSecureSMTPClient::getServerName() const {
@@ -18,6 +13,10 @@ std::string OpportunisticSecureSMTPClient::getServerName() const {
 
 unsigned int OpportunisticSecureSMTPClient::getServerPort() const {
     return jed_utils::OpportunisticSecureSMTPClient::getServerPort();
+}
+
+bool OpportunisticSecureSMTPClient::getBatchMode() const {
+    return jed_utils::OpportunisticSecureSMTPClient::getBatchMode();
 }
 
 unsigned int OpportunisticSecureSMTPClient::getCommandTimeout() const {
@@ -32,6 +31,10 @@ const Credential *OpportunisticSecureSMTPClient::getCredentials() const {
     return mCredential;
 }
 
+bool OpportunisticSecureSMTPClient::getAcceptSelfSignedCert() const {
+    return jed_utils::SecureSMTPClientBase::getAcceptSelfSignedCert();
+}
+
 void OpportunisticSecureSMTPClient::setServerName(const std::string &pServerName) {
     jed_utils::OpportunisticSecureSMTPClient::setServerName(pServerName.c_str());
 }
@@ -40,15 +43,24 @@ void OpportunisticSecureSMTPClient::setServerPort(unsigned int pPort) {
     jed_utils::OpportunisticSecureSMTPClient::setServerPort(pPort);
 }
 
+void OpportunisticSecureSMTPClient::setBatchMode(bool pEnabled) {
+    jed_utils::OpportunisticSecureSMTPClient::setBatchMode(pEnabled);
+}
+
 void OpportunisticSecureSMTPClient::setCommandTimeout(unsigned int pTimeOutInSeconds) {
     jed_utils::SMTPClientBase::setCommandTimeout(pTimeOutInSeconds);
 }
 
 void OpportunisticSecureSMTPClient::setCredentials(const Credential &pCredential) {
     jed_utils::SMTPClientBase::setCredentials(jed_utils::Credential(pCredential.getUsername().c_str(),
-                                                                    pCredential.getPassword().c_str()));
+                                                                    pCredential.getPassword().c_str(),
+                                                                    pCredential.getRecommendedAuthOption()));
     delete mCredential;
     mCredential = new Credential(pCredential);
+}
+
+void OpportunisticSecureSMTPClient::setAcceptSelfSignedCert(bool pValue) {
+    jed_utils::SecureSMTPClientBase::setAcceptSelfSignedCert(pValue);
 }
 
 void OpportunisticSecureSMTPClient::setKeepUsingBaseSendCommands(bool pValue) {