+++ src/cpp/smtpclient.cpp
@@ -1,15 +1,10 @@
 #include "smtpclient.hpp"
 #include <string>
-#include <utility>
 
 using namespace jed_utils::cpp;
 
 SmtpClient::SmtpClient(const std::string &pServerName, unsigned int pPort)
     : jed_utils::SmtpClient(pServerName.c_str(), pPort) {
-}
-
-SmtpClient::~SmtpClient() {
-    delete mCredential;
 }
 
 std::string SmtpClient::getServerName() const {
@@ -18,6 +13,10 @@ std::string SmtpClient::getServerName() const {
 
 unsigned int SmtpClient::getServerPort() const {
     return jed_utils::SmtpClient::getServerPort();
+}
+
+bool SmtpClient::getBatchMode() const {
+    return jed_utils::SmtpClient::getBatchMode();
 }
 
 unsigned int SmtpClient::getCommandTimeout() const {
@@ -38,6 +37,10 @@ void SmtpClient::setServerName(const std::string &pServerName) {
 
 void SmtpClient::setServerPort(unsigned int pPort) {
     jed_utils::SmtpClient::setServerPort(pPort);
+}
+
+void SmtpClient::setBatchMode(bool pEnabled) {
+    jed_utils::SmtpClient::setBatchMode(pEnabled);
 }
 
 void SmtpClient::setCommandTimeout(unsigned int pTimeOutInSeconds) {